New Debug screen where the functionality to send the email with the local database and the opening of the FLEX debugger is included.
